Enclosed tons of classes within "v1" namespace to indicate they're from Ogre 1.x and:  * Will eventually be ported to 2.x  * Cannot be mixed in the same RenderQueue ID with 2.x objects.

+++ OgreMain/include/OgreParticleSystem.h
@@ -529,7 +529,7 @@ namespace Ogre {
         /** @copydoc MovableObject::setRenderQueueGroup */
         void setRenderQueueGroup(uint8 queueID);
         /** @copydoc MovableObject::setRenderQueueGroupAndPriority */
-        void setRenderQueueGroupAndPriority(uint8 queueID, ushort priority);
+        void setRenderQueueGroupAndPriority(uint8 queueID, uint8 priority);
 
         /** Set whether or not particles are sorted according to the camera.
         @remarks
