+++ OgreMain/include/OgreParticleSystemRenderer.h
@@ -112,7 +112,7 @@ namespace Ogre {
         /** Sets which render queue group and priority this renderer should target with it's
             output.
         */
-        virtual void setRenderQueueGroupAndPriority(uint8 queueID, ushort priority) = 0;
+        virtual void setRenderQueueGroupAndPriority(uint8 queueID, uint8 priority) = 0;
 
         /** Setting carried over from ParticleSystem.
         */
